refactorizare
