+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/configs/CorsConfig.java
@@ -1,6 +1,5 @@
-package com.fiiconnect.api.config;
+package com.fiiconnect.api.auth_userMgmt.configs;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
