+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/models/User.java
@@ -1,9 +1,8 @@
-package com.fiiconnect.api.auth.model;
+package com.fiiconnect.api.auth_userMgmt.models;
 
-import com.fiiconnect.api.permissions.model.Role;
+import com.fiiconnect.api.auth_userMgmt.permissions.model.Role;
 import jakarta.persistence.*;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 @Entity
