+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/permissions/repository/ObiectExempluRepository.java
@@ -1,6 +1,6 @@
-package com.fiiconnect.api.permissions.repository;
+package com.fiiconnect.api.auth_userMgmt.permissions.repository;
 
-import com.fiiconnect.api.permissions.model.ObiectExemplu;
+import com.fiiconnect.api.auth_userMgmt.permissions.model.ObiectExemplu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ObiectExempluRepository extends JpaRepository<ObiectExemplu, Integer> {
