+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/permissions/repository/PermissionRepository.java
@@ -1,6 +1,6 @@
-package com.fiiconnect.api.permissions.repository;
+package com.fiiconnect.api.auth_userMgmt.permissions.repository;
 
-import com.fiiconnect.api.permissions.model.Permission;
+import com.fiiconnect.api.auth_userMgmt.permissions.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PermissionRepository extends JpaRepository<Permission, Long> {}
