+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/permissions/repository/RoleRepository.java
@@ -1,6 +1,6 @@
-package com.fiiconnect.api.permissions.repository;
+package com.fiiconnect.api.auth_userMgmt.permissions.repository;
 
-import com.fiiconnect.api.permissions.model.Role;
+import com.fiiconnect.api.auth_userMgmt.permissions.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {}
