+++ backend/api/src/main/java/com/fiiconnect/api/auth_userMgmt/validators/PasswordValidator.java
@@ -1,4 +1,4 @@
-package com.fiiconnect.api.auth.validator;
+package com.fiiconnect.api.auth_userMgmt.validators;
 
 public class PasswordValidator {
